Move shader code into separate file
